Strucutred the test framework to allow multiple suites
