+++ test/test.html
@@ -10,7 +10,7 @@
       QUnit.config.autostart = false;
   </script>
 
-  <script data-main="suite.js" src="../thirdparty/requirejs/require.js"></script>
+  <script data-main="main.js" src="../thirdparty/requirejs/require.js"></script>
 </head>
 
 <body>
