--- test/test.html~
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="utf-8">
-  <title>PumaScript Test Suite</title>
-  <link rel="stylesheet" href="qunit/qunit.css">
-  <script data-main="../dist/pumascript.js" src="../thirdparty/requirejs/require.js"></script>
-</head>
-
-<body>
-  <div id="qunit"></div>
-  <div id="qunit-fixture"></div>
-  <script src="qunit/qunit.js"></script>
-  <script src="test.js"></script>
-</body>
-</html>
